+++ src/main/java/beneficiary/Beneficiary.java
@@ -2,7 +2,7 @@ package beneficiary;
 
 public class Beneficiary {
 	private String beneficiaryName;
-	private int accNo;
+	private long accNo;
 	private String iFSCCode;
 	public String getBeneficiaryName() {
 		return beneficiaryName;
@@ -10,10 +10,10 @@ public class Beneficiary {
 	public void setBeneficiaryName(String beneficiaryName) {
 		this.beneficiaryName = beneficiaryName;
 	}
-	public int getAccNo() {
+	public long getAccNo() {
 		return accNo;
 	}
-	public void setAccNo(int accNo) {
+	public void setAccNo(long accNo) {
 		this.accNo = accNo;
 	}
 	public String getiFSCCode() {
