+++ src/main/java/beneficiary/BeneficiaryDAO.java
@@ -5,6 +5,6 @@ import java.util.List;
 public interface BeneficiaryDAO {
 	public void addBeneficiary(Beneficiary beneficiary) ;
 	public List<Beneficiary> displayBeneficiary();
-	public void updateBeneficiary(String beneficiaryName,int accNo) ;
-	public void deleteBeneficiary(int accNo) ;
+	public void updateBeneficiary(String beneficiaryName,long accNo) ;
+	public void deleteBeneficiary(long accNo) ;
 }
